+++ packages/moonstone/CHANGELOG.md
@@ -6,9 +6,9 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 ### Changed
 
-- `moonstone/ExpandableSpotlightDecorator` to focus the selected index as the default focus target.
+- `moonstone/ExpandableSpotlightDecorator` to add `getSelectedNode` variable from config to focus the selected index as the default focus.
 - `moonstone/ExpandableList` to add HOC prameter for finding selcted node.
-- `moonstone/Expandable` to add HOC prop.
+- `moonstone/Expandable` to add HOC prop. If `getSelectedNode` is defined in config, it will be used for finding selected node in `ExpandableSpotlightDecorator`.
 
 ## [2.0.0-alpha.4] - 2018-02-13